requirements.txt
django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
pytest>=8.0
pytest-django>=4.5
